+++ src/main/java/io/kontur/layers/dto/Collection.java
@@ -5,28 +5,38 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.kontur.layers.repository.model.LayersCategoryProperties;
 import io.kontur.layers.repository.model.LayersGroupProperties;
 import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Builder;
+import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
+@Data
+@Builder
 public class Collection {
 
     @JsonProperty("id")
+    @Schema(example = "address", required = true, description = "identifier of the collection used, for example, in URIs")
+    @NotNull
     private String id;
 
     @JsonProperty("title")
+    @Schema(example = "address", description = "human readable title of the collection")
     private String title;
 
     @JsonProperty("description")
+    @Schema(example = "An address.", description = "a description of the features in the collection")
     private String description;
 
     @JsonProperty("copyrights")
     private String copyrights;
 
     @JsonProperty("properties")
-    private Object properties;
+    private ObjectNode properties;
+
+    @JsonProperty("featureProperties")
+    private ObjectNode featureProperties;
 
     @JsonProperty("legend")
     private ObjectNode legend;
@@ -38,223 +48,21 @@ public class Collection {
     private LayersCategoryProperties category;
 
     @JsonProperty("links")
+    @Schema(example = "[{\"href\":\"http://data.example.com/buildings\",\"rel\":\"item\"},{\"href\":\"http://example.com/concepts/buildings.html\",\"rel\":\"describedBy\",\"type\":\"text/html\"}]", required = true, description = "")
+    @NotNull
     private List<Link> links = new ArrayList<>();
 
     @JsonProperty("extent")
     private Extent extent;
 
     @JsonProperty("itemType")
+    @Schema(description = "indicator about the type of the items in the collection (the default value is 'feature').")
     private String itemType = "feature";
 
     @JsonProperty("crs")
+    @Schema(example = "[\"http://www.opengis.net/def/crs/OGC/1.3/CRS84\",\"http://www.opengis.net/def/crs/EPSG/0/4326\"]", description = "the list of coordinate reference systems supported by the service")
     private List<String> crs;
 
-    public Collection id(String id) {
-        this.id = id;
-        return this;
-    }
-
-    @JsonProperty("id")
-    @Schema(example = "address", required = true, description = "identifier of the collection used, for example, in URIs")
-    @NotNull
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public Collection title(String title) {
-        this.title = title;
-        return this;
-    }
-
-    @JsonProperty("title")
-    @Schema(example = "address", description = "human readable title of the collection")
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public Collection description(String description) {
-        this.description = description;
-        return this;
-    }
-
-    @JsonProperty("description")
-    @Schema(example = "An address.", description = "a description of the features in the collection")
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Collection copyrights(String copyrights) {
-        this.copyrights = copyrights;
-        return this;
-    }
-
-    public String getCopyrights() {
-        return copyrights;
-    }
-
-    public void setCopyrights(String copyrights) {
-        this.copyrights = copyrights;
-    }
-
-    public Collection properties(Object properties) {
-        this.properties = properties;
-        return this;
-    }
-
-    public Object getProperties() {
-        return properties;
-    }
-
-    public void setProperties(Object properties) {
-        this.properties = properties;
-    }
-
-    public Collection legend(ObjectNode legend) {
-        this.legend = legend;
-        return this;
-    }
-
-    public ObjectNode getLegend() {
-        return legend;
-    }
-
-    public void setLegend(ObjectNode legend) {
-        this.legend = legend;
-    }
-
-    public Collection group(LayersGroupProperties group) {
-        this.group = group;
-        return this;
-    }
-
-    public LayersGroupProperties getGroup() {
-        return group;
-    }
-
-    public void setGroup(LayersGroupProperties group) {
-        this.group = group;
-    }
-
-    public Collection category(LayersCategoryProperties category) {
-        this.category = category;
-        return this;
-    }
-
-    public LayersCategoryProperties getCategory() {
-        return category;
-    }
-
-    public void setCategory(LayersCategoryProperties category) {
-        this.category = category;
-    }
-
-    public Collection links(List<Link> links) {
-        this.links = links;
-        return this;
-    }
-
-    public Collection addLinksItem(Link linksItem) {
-        this.links.add(linksItem);
-        return this;
-    }
-
-    @JsonProperty("links")
-    @Schema(example = "[{\"href\":\"http://data.example.com/buildings\",\"rel\":\"item\"},{\"href\":\"http://example.com/concepts/buildings.html\",\"rel\":\"describedBy\",\"type\":\"text/html\"}]", required = true, description = "")
-    @NotNull
-    public List<Link> getLinks() {
-        return links;
-    }
-
-    public void setLinks(List<Link> links) {
-        this.links = links;
-    }
-
-    public Collection extent(Extent extent) {
-        this.extent = extent;
-        return this;
-    }
-
-    @JsonProperty("extent")
-    @Schema(description = "")
-    public Extent getExtent() {
-        return extent;
-    }
-
-    public void setExtent(Extent extent) {
-        this.extent = extent;
-    }
-
-    public Collection itemType(String itemType) {
-        this.itemType = itemType;
-        return this;
-    }
-
-    @JsonProperty("itemType")
-    @Schema(description = "indicator about the type of the items in the collection (the default value is 'feature').")
-    public String getItemType() {
-        return itemType;
-    }
-
-    public void setItemType(String itemType) {
-        this.itemType = itemType;
-    }
-
-    public Collection crs(List<String> crs) {
-        this.crs = crs;
-        return this;
-    }
-
-    public Collection addCrsItem(String crsItem) {
-        if (this.crs == null) {
-            this.crs = new ArrayList<String>();
-        }
-        this.crs.add(crsItem);
-        return this;
-    }
-
-    @JsonProperty("crs")
-    @Schema(example = "[\"http://www.opengis.net/def/crs/OGC/1.3/CRS84\",\"http://www.opengis.net/def/crs/EPSG/0/4326\"]", description = "the list of coordinate reference systems supported by the service")
-    public List<String> getCrs() {
-        return crs;
-    }
-
-    public void setCrs(List<String> crs) {
-        this.crs = crs;
-    }
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Collection collection = (Collection) o;
-        return Objects.equals(this.id, collection.id) &&
-                Objects.equals(this.title, collection.title) &&
-                Objects.equals(this.description, collection.description) &&
-                Objects.equals(this.links, collection.links) &&
-                Objects.equals(this.extent, collection.extent) &&
-                Objects.equals(this.itemType, collection.itemType) &&
-                Objects.equals(this.crs, collection.crs);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, title, description, links, extent, itemType, crs);
-    }
+    @JsonProperty("ownedByUser")
+    private boolean ownedByUser;
 }