+++ src/main/java/io/kontur/layers/dto/CollectionCreateDto.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
 @Data
@@ -13,8 +11,6 @@ import javax.validation.constraints.Pattern;
 public class CollectionCreateDto extends CollectionUpdateDto {
 
     @JsonProperty("id")
-    @NotNull
-    @NotEmpty
     @Pattern(regexp = "[\\w]*")
     private String id;
 }