+++ src/main/java/io/kontur/layers/dto/CollectionOwner.java
@@ -1,0 +1,7 @@
+package io.kontur.layers.dto;
+
+public enum CollectionOwner {
+    ME,
+    NOT_ME,
+    ANY;
+}