+++ src/main/java/io/kontur/layers/repository/model/LayerFeature.java
@@ -3,6 +3,7 @@ package io.kontur.layers.repository.model;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.wololo.geojson.Geometry;
 
@@ -11,7 +12,8 @@ import java.time.OffsetDateTime;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class Feature {
+@EqualsAndHashCode
+public class LayerFeature {
 
     private Long layerId;
     private String featureId;